[#10] Reset to default branch if update branch is behind
Problem: if the default branch is updated while an update branch exists,
the update branch won't be automatically rebased.

Solution: in lieu of a proper rebase, which is hard, reset the update
branch to the default, as we'll have to force-push anyway, and there's
not much to rebase if there are no human commits to begin with.