[Chore] Make it possible to override input instead of updating
Problem: We'd like to have an ability to massively apply flake inputs
overrides in our repos in order to test some updates before merging
potentially breaking changes.

Solution: Make it possible to specify an attrset of input name and override
url in the list of inputs to update. For each such entry in the inputs
list, 'flake.lock' update will be done using '--override-input' instead
of '--update-input'.

Config usage example for the overriding:
```
...
    "inputs": [
        {
            "input": "nixpkgs",
            "override_url": "github:nixos/nixpkgs"
        }
    ],
    "repos": [
      {
        "owner": "serokell",
        "repo": "update-daemon",
        "settings": {},
        "type": "github"
      }
    ],
...
```