[Chore] Propely support GitLab repos during config.json generation
Problem: Currently, JSON config generation is only capable of generating
config that is compatible with 'RepoHandle::GitHub' and an attempt to
use it for GitLab repo fails.

Solution: Rework JSON config generation to separately handle GitHub and
GitLab repos from the module config and generate attrsets that are
compatible with 'RepoHandle' type.