[#45] Use `flake update` instead of `flake lock --update-input`
Problem: Newer nix versions report the following warning when nix flake
lock --update-input is used:
```
warning: '--update-input' is a deprecated alias for 'flake update' and
will be removed in a future version
```
However, update-daemon still uses this option, so it's better to switch to the newer interface.

Solution: Replace command `flake lock --update-inputs` with `flake
update` in case when a list of inputs to update is provided.